requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
opencv-python>=4.5.3.56
tqdm>=4.62.0
scikit-learn>=1.0.0
scikit-image>=0.18.3
scipy>=1.7.0
matplotlib>=3.4.3
seaborn>=0.11.2
pydicom>=2.2.2
albumentations>=1.1.0
segmentation-models-pytorch>=0.3.3
